Moved test_util tests to test/utils
