+++ backend/test/utils/test_embedding_util.py
